Update api DTO and refactor the observer to call patch endpoint
