+++ core/src/main/java/org/stellar/anchor/apiclient/PlatformApiClient.java
@@ -1,4 +1,4 @@
-package org.stellar.anchor.reference.client;
+package org.stellar.anchor.apiclient;
 
 import java.io.IOException;
 import java.util.HashMap;
