+++ core/src/main/java/org/stellar/anchor/event/EventPublisher.java
@@ -1,7 +1,7 @@
 package org.stellar.anchor.event;
 
+import org.stellar.anchor.api.event.AnchorEvent;
 import org.stellar.anchor.api.exception.EventPublishException;
-import org.stellar.anchor.event.models.AnchorEvent;
 
 public interface EventPublisher {
   void publish(String queue, AnchorEvent event) throws EventPublishException;
