+++ core/src/main/java/org/stellar/anchor/event/EventService.java
@@ -1,46 +1,14 @@
 package org.stellar.anchor.event;
 
-import static org.stellar.anchor.util.Log.errorF;
-
-import io.micrometer.core.instrument.Metrics;
-import java.util.Map;
+import org.stellar.anchor.api.event.AnchorEvent;
 import org.stellar.anchor.api.exception.EventPublishException;
-import org.stellar.anchor.config.event.EventConfig;
-import org.stellar.anchor.event.models.AnchorEvent;
+import org.stellar.anchor.sep24.Sep24Transaction;
+import org.stellar.anchor.sep31.Sep31Transaction;
 
-public class EventService {
-  private final EventConfig eventConfig;
-  private EventPublisher eventPublisher;
+public interface EventService {
+  void publish(Sep24Transaction txn, AnchorEvent.Type type);
 
-  private final Map<String, String> eventTypeMapping;
+  void publish(Sep31Transaction txn, AnchorEvent.Type type);
 
-  public EventService(EventConfig eventConfig) {
-    this.eventConfig = eventConfig;
-    this.eventTypeMapping = eventConfig.getEventTypeToQueue();
-  }
-
-  public void setEventPublisher(EventPublisher eventPublisher) {
-    this.eventPublisher = eventPublisher;
-  }
-
-  public void publish(AnchorEvent event) throws EventPublishException {
-    if (eventConfig.isEnabled()) {
-      // publish the event
-      eventPublisher.publish(getQueue(event.getType()), event);
-      // update metrics
-      Metrics.counter(
-              "event.published", "class", event.getClass().getSimpleName(), "type", event.getType())
-          .increment();
-    }
-  }
-
-  String getQueue(String eventType) {
-    String queue = eventTypeMapping.get(eventType);
-    if (queue == null) {
-      errorF("There is no queue defined for event type:{}", eventType);
-      throw new RuntimeException(
-          String.format("There is no queue defined for event type:%s", eventType));
-    }
-    return queue;
-  }
+  void publish(AnchorEvent event) throws EventPublishException;
 }
