--- core/src/main/java/org/stellar/anchor/event/models/AnchorEvent.java
@@ -1,10 +1,0 @@
-package org.stellar.anchor.event.models;
-
-public interface AnchorEvent {
-  String TYPE_TRANSACTION = "TransactionEvent";
-  String TYPE_QUOTE = "QuoteEvent";
-
-  String getType();
-
-  String getEventId();
-}
