+++ platform/src/main/java/org/stellar/anchor/platform/event/NoopEventPublisher.java
@@ -1,9 +1,7 @@
 package org.stellar.anchor.platform.event;
 
-import static org.stellar.anchor.util.Log.debugF;
-
+import org.stellar.anchor.api.event.AnchorEvent;
 import org.stellar.anchor.event.EventPublisher;
-import org.stellar.anchor.event.models.AnchorEvent;
 
 public class NoopEventPublisher implements EventPublisher {
   @Override
