+++ platform/src/main/java/org/stellar/anchor/platform/event/SqsEventPublisher.java
@@ -6,9 +6,9 @@ import com.amazonaws.services.sqs.model.MessageAttributeValue;
 import com.amazonaws.services.sqs.model.SendMessageRequest;
 import com.amazonaws.services.sqs.model.SendMessageResult;
 import com.google.gson.Gson;
+import org.stellar.anchor.api.event.AnchorEvent;
 import org.stellar.anchor.api.exception.EventPublishException;
 import org.stellar.anchor.event.EventPublisher;
-import org.stellar.anchor.event.models.AnchorEvent;
 import org.stellar.anchor.platform.config.SqsConfig;
 import org.stellar.anchor.util.Log;
