+++ platform/src/main/java/org/stellar/anchor/platform/service/TransactionService.java
@@ -1,7 +1,7 @@
 package org.stellar.anchor.platform.service;
 
+import static org.stellar.anchor.api.event.AnchorEvent.Type.TRANSACTION_STATUS_CHANGED;
 import static org.stellar.anchor.api.sep.SepTransactionStatus.*;
-import static org.stellar.anchor.event.models.TransactionEvent.Type.TRANSACTION_STATUS_CHANGED;
 import static org.stellar.anchor.sep31.Sep31Helper.allAmountAvailable;
 import static org.stellar.anchor.util.MathHelper.decimal;
 import static org.stellar.anchor.util.MathHelper.equalsAsDecimals;
@@ -19,11 +19,8 @@ import org.stellar.anchor.api.platform.*;
 import org.stellar.anchor.api.sep.AssetInfo;
 import org.stellar.anchor.api.sep.SepTransactionStatus;
 import org.stellar.anchor.api.shared.Amount;
-import org.stellar.anchor.api.shared.RefundPayment;
-import org.stellar.anchor.api.shared.Refunds;
 import org.stellar.anchor.asset.AssetService;
 import org.stellar.anchor.event.EventService;
-import org.stellar.anchor.event.models.TransactionEvent;
 import org.stellar.anchor.platform.data.JdbcSep24Transaction;
 import org.stellar.anchor.platform.data.JdbcSep31Transaction;
 import org.stellar.anchor.platform.data.JdbcSepTransaction;
@@ -330,119 +327,5 @@ public class TransactionService {
         throw new BadRequestException("amount_fee != sum(quote.fee.total)");
       }
     }
-  }
-
-  GetTransactionResponse toGetTransactionResponse(JdbcSepTransaction txn) throws SepException {
-    switch (txn.getProtocol()) {
-      case "24":
-        return toGetTransactionResponse((JdbcSep24Transaction) txn);
-      case "31":
-        return toGetTransactionResponse((JdbcSep31Transaction) txn);
-      default:
-        throw new SepException(String.format("Unsupported protocol:%s", txn.getProtocol()));
-    }
-  }
-
-  RefundPayment toRefundPayment(
-      org.stellar.anchor.sep31.RefundPayment refundPayment, String assetName) {
-    return RefundPayment.builder()
-        .id(refundPayment.getId())
-        .idType(RefundPayment.IdType.STELLAR)
-        .amount(new Amount(refundPayment.getAmount(), assetName))
-        .fee(new Amount(refundPayment.getFee(), assetName))
-        .requestedAt(null)
-        .refundedAt(null)
-        .build();
-  }
-
-  Refunds toRefunds(Sep31Refunds refunds, String assetName) {
-    // build payments
-    RefundPayment[] payments =
-        refunds.getRefundPayments().stream()
-            .map(refundPayment -> toRefundPayment(refundPayment, assetName))
-            .toArray(RefundPayment[]::new);
-
-    return Refunds.builder()
-        .amountRefunded(new Amount(refunds.getAmountRefunded(), assetName))
-        .amountFee(new Amount(refunds.getAmountFee(), assetName))
-        .payments(payments)
-        .build();
-  }
-
-  private GetTransactionResponse toGetTransactionResponse(JdbcSep31Transaction txn) {
-    Refunds refunds = null;
-    if (txn.getRefunds() != null) {
-      refunds = toRefunds(txn.getRefunds(), txn.getAmountInAsset());
-    }
-
-    return GetTransactionResponse.builder()
-        .id(txn.getId())
-        .sep(31)
-        .kind(TransactionEvent.Kind.RECEIVE.getKind())
-        .status(txn.getStatus())
-        .amountExpected(new Amount(txn.getAmountExpected(), txn.getAmountInAsset()))
-        .amountIn(new Amount(txn.getAmountIn(), txn.getAmountInAsset()))
-        .amountOut(new Amount(txn.getAmountOut(), txn.getAmountOutAsset()))
-        .amountFee(new Amount(txn.getAmountFee(), txn.getAmountFeeAsset()))
-        .quoteId(txn.getQuoteId())
-        .startedAt(txn.getStartedAt())
-        .updatedAt(txn.getUpdatedAt())
-        .completedAt(txn.getCompletedAt())
-        .transferReceivedAt(txn.getTransferReceivedAt())
-        .message(txn.getRequiredInfoMessage()) // Assuming these are meant to be the same.
-        .refunds(refunds)
-        .stellarTransactions(txn.getStellarTransactions())
-        .externalTransactionId(txn.getExternalTransactionId())
-        .customers(txn.getCustomers())
-        .creator(txn.getCreator())
-        .build();
-  }
-
-  RefundPayment toRefundPayment(Sep24RefundPayment refundPayment, String assetName) {
-    return RefundPayment.builder()
-        .id(refundPayment.getId())
-        .idType(RefundPayment.IdType.STELLAR)
-        .amount(new Amount(refundPayment.getAmount(), assetName))
-        .fee(new Amount(refundPayment.getFee(), assetName))
-        .requestedAt(null)
-        .refundedAt(null)
-        .build();
-  }
-
-  Refunds toRefunds(Sep24Refunds refunds, String assetName) {
-    // build payments
-    RefundPayment[] payments =
-        refunds.getRefundPayments().stream()
-            .map(refundPayment -> toRefundPayment(refundPayment, assetName))
-            .toArray(RefundPayment[]::new);
-
-    return Refunds.builder()
-        .amountRefunded(new Amount(refunds.getAmountRefunded(), assetName))
-        .amountFee(new Amount(refunds.getAmountFee(), assetName))
-        .payments(payments)
-        .build();
-  }
-
-  private GetTransactionResponse toGetTransactionResponse(JdbcSep24Transaction txn) {
-    Refunds refunds = null;
-    if (txn.getRefunds() != null) {
-      refunds = toRefunds(txn.getRefunds(), txn.getAmountInAsset());
-    }
-
-    return GetTransactionResponse.builder()
-        .id(txn.getId())
-        .sep(24)
-        .kind(txn.getKind())
-        .status(txn.getStatus())
-        .amountIn(new Amount(txn.getAmountIn(), txn.getAmountInAsset()))
-        .amountOut(new Amount(txn.getAmountOut(), txn.getAmountOutAsset()))
-        .amountFee(new Amount(txn.getAmountFee(), txn.getAmountFeeAsset()))
-        .startedAt(txn.getStartedAt())
-        .updatedAt(txn.getUpdatedAt())
-        .completedAt(txn.getCompletedAt())
-        .refunds(refunds)
-        .stellarTransactions(txn.getStellarTransactions())
-        .externalTransactionId(txn.getExternalTransactionId())
-        .build();
   }
 }
